Microsoft_Purview_Utility/requirements.txt
pandas
requests
azure-identity
azure-purview-catalog
azure-purview-datamap
azure-core
python-dotenv
flask
flask-cors


